+++ cmd/privnet_create.go
@@ -1,13 +1,11 @@
 package cmd
 
 import (
-	"bufio"
+	"fmt"
 	"net"
-	"os"
+	"strconv"
 
-	"github.com/exoscale/cli/table"
 	"github.com/exoscale/egoscale"
-
 	"github.com/spf13/cobra"
 )
 
@@ -25,18 +23,42 @@ var privnetCreateCmd = &cobra.Command{
 		if err != nil {
 			return err
 		}
-		startip, err := cmd.Flags().GetString("startip")
+		sip, err := cmd.Flags().GetString("startip")
 		if err != nil {
 			return err
 		}
-		endip, err := cmd.Flags().GetString("endip")
+		startip := net.ParseIP(sip)
+
+		eip, err := cmd.Flags().GetString("endip")
 		if err != nil {
 			return err
 		}
-		netmask, err := cmd.Flags().GetString("netmask")
+		endip := net.ParseIP(eip)
+
+		nmask, err := cmd.Flags().GetString("netmask")
 		if err != nil {
 			return err
 		}
+
+		cidrmask, err := cmd.Flags().GetString("cidrmask")
+		if err != nil {
+			return err
+		}
+
+		if nmask != "" && cidrmask != "" {
+			return fmt.Errorf("netmask %q and cidrmask %q are mutually exclusive", nmask, cidrmask)
+		}
+
+		netmask := net.ParseIP(nmask)
+		if netmask == nil && cidrmask != "" {
+			c, err := strconv.ParseInt(cidrmask, 10, 32)
+			if err != nil {
+				return err
+			}
+			ipmask := net.CIDRMask(int(c), 32)
+			netmask = (net.IP)(ipmask)
+		}
+
 		zone, err := cmd.Flags().GetString("zone")
 		if err != nil {
 			return err
@@ -46,27 +68,16 @@ var privnetCreateCmd = &cobra.Command{
 			zone = gCurrentAccount.DefaultZone
 		}
 
-		if zone == "" {
-			reader := bufio.NewReader(os.Stdin)
-			if desc == "" {
-				desc, err = readInput(reader, "Description", "")
-				if err != nil {
-					return err
-				}
-			}
-			if zone == "" {
-				zone, err = readInput(reader, "Zone", gCurrentAccount.DefaultZone)
-				if err != nil {
-					return err
-				}
-			}
-		}
-
 		if isEmptyArgs(name, zone) {
 			return cmd.Usage()
 		}
 
-		return privnetCreate(name, desc, zone, startip, endip, netmask)
+		newNet, err := privnetCreate(name, desc, zone, startip, endip, netmask)
+		if err != nil {
+			return err
+		}
+
+		return privnetShow(*newNet)
 	},
 }
 
@@ -79,53 +90,50 @@ func isEmptyArgs(args ...string) bool {
 	return false
 }
 
-func privnetCreate(name, desc, zoneName, startIPAddr, endIPAddr, netmask string) error {
-	zone, err := getZoneIDByName(zoneName)
+func privnetCreate(name, desc, zoneName string, startIP, endIP, netmask net.IP) (*egoscale.Network, error) {
+	zoneID, err := getZoneIDByName(zoneName)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	resp, err := cs.RequestWithContext(gContext, &egoscale.ListNetworkOfferings{ZoneID: zone, Name: "PrivNet"})
-	if err != nil {
-		return err
+	// NetworkOffering are cross zones
+	listReq := &egoscale.ListNetworkOfferings{
+		Name:     "PrivNet",
+		ZoneID:   zoneID,
+		Page:     1,
+		PageSize: 1,
 	}
 
-	startIP := net.ParseIP(startIPAddr)
-	endIP := net.ParseIP(endIPAddr)
-	netmaskIP := net.ParseIP(netmask)
+	resp, err := cs.RequestWithContext(gContext, listReq)
+	if err != nil {
+		return nil, err
+	}
 
-	s := resp.(*egoscale.ListNetworkOfferingsResponse)
+	nos := resp.(*egoscale.ListNetworkOfferingsResponse)
+	if len(nos.NetworkOffering) != 1 {
+		return nil, fmt.Errorf("missing Network Offering %q in %q", listReq.Name, zoneName)
+	}
+
+	if startIP != nil && endIP != nil && netmask == nil {
+		netmask = net.IPv4(255, 255, 255, 0)
+	}
 
 	req := &egoscale.CreateNetwork{
-		DisplayText: desc,
-		Name:        name,
-		ZoneID:      zone,
-		StartIP:     startIP,
-		EndIP:       endIP,
-		Netmask:     netmaskIP,
-	}
-	if len(s.NetworkOffering) > 0 {
-		req.NetworkOfferingID = s.NetworkOffering[0].ID
+		Name:              name,
+		DisplayText:       desc,
+		ZoneID:            zoneID,
+		StartIP:           startIP,
+		EndIP:             endIP,
+		Netmask:           netmask,
+		NetworkOfferingID: nos.NetworkOffering[0].ID,
 	}
 
 	resp, err = cs.RequestWithContext(gContext, req)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	newNet := resp.(*egoscale.Network)
-
-	table := table.NewTable(os.Stdout)
-	table.SetHeader([]string{"Name", "Description", "ID", "DHCP"})
-
-	table.Append([]string{
-		newNet.Name,
-		newNet.DisplayText,
-		newNet.ID.String(),
-		dhcpRange(startIP, endIP, netmaskIP),
-	})
-	table.Render()
-	return nil
+	return resp.(*egoscale.Network), nil
 }
 
 func init() {
@@ -133,6 +141,7 @@ func init() {
 	privnetCreateCmd.Flags().StringP("startip", "s", "", "the beginning IP address in the network IP range. Required for managed networks.")
 	privnetCreateCmd.Flags().StringP("endip", "e", "", "the ending IP address in the network IP range. Required for managed networks.")
 	privnetCreateCmd.Flags().StringP("netmask", "n", "", "the netmask of the network. Required for managed networks.")
+	privnetCreateCmd.Flags().StringP("cidrmask", "c", "", "the cidrmask of the network. Required for managed networks.")
 	privnetCreateCmd.Flags().StringP("zone", "z", "", "Assign private network to a zone")
 	privnetCmd.AddCommand(privnetCreateCmd)
 }