+++ cmd/privnet_show.go
@@ -30,7 +30,7 @@ var privnetShowCmd = &cobra.Command{
 		}
 
 		w := tabwriter.NewWriter(os.Stdout, 0, 0, 1, ' ', tabwriter.TabIndent)
-		dhcp := dhcpRange(network.StartIP, network.EndIP, network.Netmask)
+		dhcp := dhcpRange(*network)
 
 		fmt.Fprintf(w, "Network:\t%s\n", network.Name)            // nolint: errcheck
 		fmt.Fprintf(w, "Description:\t%s\n", network.DisplayText) // nolint: errcheck