+++ cmd/privnet_test.go
@@ -3,48 +3,40 @@ package cmd
 import (
 	"net"
 	"testing"
-)
 
-type dhcp struct {
-	startIP net.IP
-	endIP   net.IP
-	netmask net.IP
-}
+	"github.com/exoscale/egoscale"
+)
 
 func TestDhcpRange(t *testing.T) {
 	cases := []struct {
-		in   dhcp
-		want string
+		network  egoscale.Network
+		expected string
 	}{
 		{
-			in: dhcp{
-				startIP: net.ParseIP("10.0.0.2"),
-				endIP:   net.ParseIP("10.0.0.100"),
-				netmask: net.ParseIP("255.255.255.0"),
+			egoscale.Network{
+				StartIP: net.IPv4(10, 0, 0, 2),
+				EndIP:   net.IPv4(10, 0, 0, 100),
+				Netmask: net.IPv4(255, 255, 255, 0),
 			},
-			want: "10.0.0.2-10.0.0.100 /24",
+			"10.0.0.2-10.0.0.100 /24",
 		},
 		{
-			in: dhcp{
-				startIP: net.ParseIP("10.0.0.1"),
-				endIP:   net.ParseIP("10.0.0.200"),
-				netmask: net.ParseIP("255.255.0.0"),
+			egoscale.Network{
+				StartIP: net.IPv4(10, 0, 0, 1),
+				EndIP:   net.IPv4(10, 0, 0, 200),
+				Netmask: net.IPv4(255, 255, 0, 0),
 			},
-			want: "10.0.0.1-10.0.0.200 /16",
+			"10.0.0.1-10.0.0.200 /16",
 		},
 		{
-			in: dhcp{
-				startIP: net.ParseIP("10.0.0.1"),
-				endIP:   net.ParseIP("foo"),
-				netmask: net.ParseIP("255.255.0.0"),
-			},
-			want: "n/a",
+			egoscale.Network{},
+			"n/a",
 		},
 	}
-	for _, c := range cases {
-		result := dhcpRange(c.in.startIP, c.in.endIP, c.in.netmask)
-		if result != c.want {
-			t.Errorf("got %s, want %s", result, c.want)
+	for _, tt := range cases {
+		result := dhcpRange(tt.network)
+		if result != tt.expected {
+			t.Errorf("got %s, want %s", result, tt.expected)
 		}
 	}
 }